requirements.txt
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==14.0
croniter==2.0.5
cryptography==42.0.8
dagster==1.7.9
dagster-cloud==1.7.9
dagster-cloud-cli==1.7.9
dagster-graphql==1.7.9
dagster-pipes==1.7.9
dagster-webserver==1.7.9
docstring_parser==0.16
exceptiongroup==1.2.1
filelock==3.14.0
Flask==3.0.3
fsspec==2024.6.0
github3.py==4.0.1
gql==3.5.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
grpcio==1.64.1
grpcio-health-checking==1.62.2
h11==0.14.0
httptools==0.6.1
humanfriendly==10.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
packaging==24.0
pendulum==3.0.0
pex==2.4.0
pluggy==1.5.0
polars==0.20.31
prompt-toolkit==3.0.36
protobuf==4.25.3
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
questionary==2.0.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
starlette==0.37.2
structlog==24.2.0
tabulate==0.9.0
time-machine==2.14.1
tomli==2.0.1
toposort==1.10
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
universal_pathlib==0.2.2
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
yarl==1.9.4
